tms9995: add some initial sketch PIC support
This is well beyond what the linker can handle right now but with a bit of
tooling we should be able to generate Amiga style shared libraries or similar
where R15 is the base of the library static data. The static data will need
some kind of relocation handling to be generalized, and lib->lib calls mean
we'd have to find somewhere to put the old R15 each time, but it ought to
be doable.

While we can't do it until the reg maps get fixed it would suffice to make
PIC using library entry routines save R15 as if it were a register variable.
A bit of overhead but not too bad.