tms9995: special case 32bit load/stores
The compiler core really can't cope with the idea of having a pointer register
that simply isn't usable as a zero index it seems unless its a different class
and different classes mess up moves because of the move limitation.

It will turn all the @n(r0) into ai n,r0 *r0 but that doesn't help for 32bit
sized data.

So instead spot the r0 cases and fix them up to inct/dect for 32bit access.
Ugly but seems to be a viable path and the cost is pretty low.