tms9995: table fixes
Fix wrong regtype on convert of int to char. This caused the object to move
register but the generated code did not. Now we use RLEFT so it stays put.

Deal with function calls. This is ugly but will do for now

On the TI990/TMS99xx B is one less level of indirection than most processors
so
	B R0

actually begins executing the code in register 0, wherebas B *R0 executes the
code from PC = R0. The TI990/12 has BIND (branch indirect) but we don't so
isntead generate move into R1 in these cases and then use *R1. We know R1
will be safe because the branch itself will trash it on return.

Just remember not to use it for a register argument in future.