tms9995: add argument promotion for untyped
This is really a flaw in the base compiler - it doesn't do this work for us.

We do it by always pushing char/unsigned char as 16bit types as we need to
do for untyped, and fudging the stack offset by 1 the other side if typed.

Need to work out how to spot typed/untyped properly and then it can use a
slightly simpler and faster method.