tms9995: clean up vararg handling
Still not entirely right. We break on the case of

	int x(char *p, ....)

	p("%ld", 0L);

because the long isn't half loaded into a register and half stack.

We either need to do that ugly (as parisc seems to) or better yet work out
how to tell funcode is being called for a vararg typed target and not generate
anything but pushes